Example Wcf Service
This web service has an endpoint for SOAP and an endpoint for REST. The
SOAP endpoint has one version of each method and it only accepts Xml and
returns Xml. The REST endpoint has two versions of each method that
accept Xml or Json but one version only returns Xml and the other only
returns Json. Use this web service and the accompanying Ajax examples to
help you learn how to implement SOAP and REST endpoints in your own web
service. Pen Blade LLC assumes no liability for use of this code.

This project is limited in scope to providing a few examples on how to
setup a WCF web service with REST (REpresentational State Transfer)
endpoint to submit Xml or Json and return Xml or Json while also
providing a standard SOAP (Simple Object Access Protocol) endpoint.